Added documentation source and output based on readme.md
The documentation is created from the readme.md file and uses docnet
v0.13 or higher to generate its output. Documentation\UserGuideSource is
the folder which contains the markdown. Run generatehtml.cmd which will
generate html files from the md files into documentation\userguidehtml.
I committed the html output as well so there's an initial batch.
Configure where the files are generated in
documentation\userguidesource\docnet.json as well as other things
related to the html generation process.

For #219