Added Api docs generation directives
Uses Sandcastle Help File Builder: https://github.com/EWSoftware/SHFB
This has to be installed.
It generates an API reference manual similar to MSDN, including search.
This is configurable through the GUI of SHFB, if you want to change
anything, or alter the shfbproj file in a text editor.
Generation uses msbuild so vsvars32.bat is called to get msbuild into
the path. If this isn't needed, remove it from the generatehtml.cmd file
It generates output into docs\apihtml which is ignored in the gitignore
file.