Add LINQPad logging (#903)
Adds a logger that sends coloured output to LINQPad, fixing #447.

This new logger is enabled by default (in `DefaultConfig`) if the benchmark detects that it's being run under LINQPad. This makes it really easy to use, but I wasn't sure if this "magic" was appropriate; let me know if you'd like it to be done a different way.