[wasm] Add a `--wasmDataDir` parameter (#1938)
In an earlier commit, generated projects got support for importing
props/targets files from the benchmark projects. And that allows
customizing the build.

But there isn't a clean way to inject/add references to the build of the
project, like custom paths, or access to additional files, without
adding new command line parameters.

To solve that, `--wasmDataDir` param is added which is surfaced in the
generated project as the msbuild property - `$(WasmDataDir)`.

An example use case is to have the wasm js file added to this data dir,
and in the props file use:
`<WasmMainJSPath>$(WasmDataDir)test-main.js</WasmMainJSPath>`.