DotNetCliCommand: Add new `RetryFailedBuildWithNoDeps` property
If a build fails, then it is retried with `--no-dependencies`. But this
can be problematic because msbuild can end up skipping some steps,
resulting in an incorrect build. And that can cause inexplicable issues
when running the benchmarks.

This is seen with wasm/aot builds with emcc, or wasm-ld crashes.

The property defaults to true, but is disabled for wasm builds.