Bugfix MetricColumn: Respect unit when formatting values.
The previous implementation would only use the given unit if SummaryStyle.PrintUnitsInContent was true, otherwise it would fallback to the default unit, regardless whether SummaryStyle.PrintUnitsInHeader was true or not. This fix takes both SummaryStyle.PrintUnitsIn* into account.

It should be noted that other columns than the MetricColumn do **not** take the visibility of units into account and always use the given unit, be it explicitly selected by the user, or calculated via GetBest*Unit().