Update changelog download strategy in workflows
The download strategies used in "docs-changelog-generate.yaml" and "docs-stable.yaml" workflow files are updated to control the number of versions retrieved for the changelog instead of getting all or latest versions. In "Program.cs", changed the condition to accept the "VersionCount" argument and implemented a loop to fetch 'n' latest versions where 'n' is the given argument value.