Add version history to BuildContext
A new property called 'VersionHistory' was added to the 'BuildContext' class. This was done to manage historical version data and facilitate the automatic processing of changelogs for different versions. The version data was previously stored in the 'DocumentationHelper' class, but this was renamed to 'VersionHistory' and refactored to fetch historical data from a text file ('versions.txt'). This addition and changes improve the manageability and flexibility of handling version data for documentation generation.